device details to devices
